refactor: implement react-query for games fetching (improves caching and reduces API calls)
