--- src/hooks/useData.ts
@@ -1,38 +1,0 @@
-import { useState, useEffect } from 'react'
-import apiClient from '../services/api-client'
-import { AxiosRequestConfig } from 'axios'
-
-export type FetchResponse<T> = {
-  count: number
-  results: T[]
-}
-
-const useData = <T>(endpoint: string, requestConfig?: AxiosRequestConfig, deps?: any[]) => {
-  const [data, setData] = useState<T[]>([])
-  const [error, setError] = useState('')
-  const [isLoading, setLoading] = useState(false)
-
-  useEffect(() => {
-    const controller = new AbortController()
-
-    setLoading(true)
-    apiClient
-      .get<FetchResponse<T>>(endpoint, { signal: controller.signal, ...requestConfig })
-      .then((res) => {
-        setData(res.data.results)
-        setLoading(false)
-      })
-      .catch((err) => {
-        if (err.name !== 'CanceledError') {
-          setError(err.message)
-          setLoading(false)
-        }
-      })
-
-    return () => controller.abort()
-  }, deps ? [...deps] : [])
-
-  return { data, error, isLoading }
-}
-
-export default useData
