+++ src/hooks/useGames.ts
@@ -1,5 +1,6 @@
+import { useQuery } from '@tanstack/react-query'
 import { type GameQuery } from '../App'
-import useData from './useData'
+import apiClient, { type FetchResponse } from '../services/api-client'
 
 export type Platform = {
   id: number
@@ -16,20 +17,20 @@ export type Game = {
   rating_top: number
 }
 
-const useGames = (
-  gameQuery: GameQuery
-) =>
-  useData<Game>(
-    '/games',
-    {
-      params: {
-        genres: gameQuery.genre?.id,
-        platforms: gameQuery.platform?.id,
-        ordering: gameQuery.sortOrder,
-        search: gameQuery.searchText
-      }
-    },
-    [gameQuery]
-  )
+const useGames = (gameQuery: GameQuery) =>
+  useQuery<FetchResponse<Game>, Error>({
+    queryKey: ['games', gameQuery],
+    queryFn: () =>
+      apiClient
+        .get<FetchResponse<Game>>('/games', {
+          params: {
+            genres: gameQuery.genre?.id,
+            parent_platforms: gameQuery.platform?.id,
+            ordering: gameQuery.sortOrder,
+            search: gameQuery.searchText
+          }
+        })
+        .then((res) => res.data)
+  })
 
 export default useGames
