+++ src/hooks/usePlatforms.ts
@@ -1,8 +1,7 @@
 import { useQuery } from '@tanstack/react-query'
 import platforms from '../data/platforms'
-import apiClient from '../services/api-client'
-import { FetchResponse } from './useData'
-import { Platform } from './useGames'
+import apiClient, { type FetchResponse } from '../services/api-client'
+import { type Platform } from './useGames'
 
 const usePlatforms = () =>
   useQuery({
