+++ src/services/api-client.ts
@@ -1,4 +1,9 @@
-import axios from "axios";
+import axios from 'axios'
+
+export type FetchResponse<T> = {
+  count: number
+  results: T[]
+}
 
 export default axios.create({
   baseURL: 'https://api.rawg.io/api',
